refactor(proxy): route-policy module + extracted helpers (JOV-2446)
- proxy-routing owns all classification + APP_ROUTES-derived reserved
- categorizePath returns publicProfileCandidate (bugfix)
- extracted investor, clerk-fapi, final-response
- new tests for candidate excludes (start/pricing/about/investors)
- suite green, typecheck+biome pass
- thin dispatcher, no Clerk change
